move launch_testing into launch_testing.legacy namespace (#130)
* move launch_testing module into legacy namespace

* add legacy imports to maintain backward compatibility
